requirements.txt
requests
google-cloud-storage
beautifulsoup4
functions-framework